python/webserverWithLogin/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
itsdangerous==2.1.2
click==8.1.7